% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{sc_promoter}
\alias{sc_promoter}
\title{Chromium Single Cell Multiome ATAC + Gene Expression assay}
\description{
An example dataset of Single cell Gene and ATAC-seq data for
Construction script is in 'inst/script/make-multiome-data.R'. Function returns a promoter GRanges
object with pseudo bulk counts stored in metadata column "counts2" as NumericList.
}
\examples{

data("sc_promoter")

}
